Le Big Push
Version 1.0!  Most of this was written in one day.  FTPhandler saw some major work after the major stint to make it less fragile (it now creates folders if they aren't there, among other things).

Updated the readme, and made a separate config file to decouple the code and make it more human-compatible.